api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
aiosqlite==0.21.0
alembic==1.15.2
pydantic-settings==2.13.1
lxml==5.3.2
python-dateutil==2.9.0
cryptography==44.0.2
apscheduler==3.11.0
httpx==0.28.1
structlog==25.5.0
tenacity==9.1.4
zeep==4.3.2
